requirements.txt
flask>=2.3.2
langchain>=0.2.1
langchain-openai>=0.1.7
langchain-community>=0.0.28
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
python-dotenv>=1.0.1
